+++ examples/src/tests/group.rs
@@ -1,6 +1,6 @@
 use autd3::{driver::link::Link, prelude::*};
 
-pub async fn group_by_device(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
+pub fn group_by_device(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
     let center = autd.center() + Vector3::new(0., 0., 150.0 * mm);
 
     autd.group(|dev| match dev.idx() {
@@ -10,13 +10,12 @@ pub async fn group_by_device(autd: &mut Controller<impl Link>) -> anyhow::Result
     })
     .set("null", (Static::new(), Null::new()))?
     .set("focus", (Sine::new(150. * Hz), Focus::new(center)))?
-    .send()
-    .await?;
+    .send()?;
 
     Ok(true)
 }
 
-pub async fn group_by_transducer(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
+pub fn group_by_transducer(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
     let cx = autd.center().x;
     let g1 = Focus::new(autd[0].center() + Vector3::new(0., 0., 150.0 * mm));
     let g2 = Null::new();
@@ -33,7 +32,7 @@ pub async fn group_by_transducer(autd: &mut Controller<impl Link>) -> anyhow::Re
     .set("null", g2)?;
 
     let m = Sine::new(150. * Hz);
-    autd.send((m, g)).await?;
+    autd.send((m, g))?;
 
     Ok(true)
 }