+++ examples/src/tests/test_runner.rs
@@ -7,43 +7,38 @@ use super::{
     stm::*, user_defined_gain_modulation::*,
 };
 
-pub async fn run<L: Link>(mut autd: Controller<L>) -> anyhow::Result<()> {
+pub fn run<L: Link>(mut autd: Controller<L>) -> anyhow::Result<()> {
     type Test<L> = (
         &'static str,
-        fn(
-            &'_ mut Controller<L>,
-        )
-            -> std::pin::Pin<Box<dyn std::future::Future<Output = anyhow::Result<bool>> + '_>>,
+        fn(&'_ mut Controller<L>) -> anyhow::Result<bool>,
     );
 
     println!("======== AUTD3 firmware information ========");
-    autd.firmware_version().await?.iter().for_each(|firm_info| {
+    autd.firmware_version()?.iter().for_each(|firm_info| {
         println!("{}", firm_info);
     });
     println!("============================================");
 
     let mut examples: Vec<Test<_>> = vec![
-        ("Single focus test", |autd| Box::pin(focus(autd))),
-        ("Bessel beam test", |autd| Box::pin(bessel(autd))),
-        ("Plane wave test", |autd| Box::pin(plane(autd))),
-        ("Wav modulation test", |autd| Box::pin(audio_file(autd))),
-        ("FociSTM test", |autd| Box::pin(foci_stm(autd))),
-        ("GainSTM test", |autd| Box::pin(gain_stm(autd))),
-        ("Multiple foci test", |autd| Box::pin(holo(autd))),
-        ("FIR test", |autd| Box::pin(fir(autd))),
+        ("Single focus test", |autd| focus(autd)),
+        ("Bessel beam test", |autd| bessel(autd)),
+        ("Plane wave test", |autd| plane(autd)),
+        ("Wav modulation test", |autd| audio_file(autd)),
+        ("FociSTM test", |autd| foci_stm(autd)),
+        ("GainSTM test", |autd| gain_stm(autd)),
+        ("Multiple foci test", |autd| holo(autd)),
+        ("FIR test", |autd| fir(autd)),
         ("User-defined Gain & Modulation test", |autd| {
-            Box::pin(user_defined(autd))
+            user_defined(autd)
         }),
-        ("Flag test", |autd| Box::pin(flag(autd))),
-        ("Custom Gain test", |autd| Box::pin(custom(autd))),
+        ("Flag test", |autd| flag(autd)),
+        ("Custom Gain test", |autd| custom(autd)),
         ("Group (by Transducer) test", |autd| {
-            Box::pin(group_by_transducer(autd))
+            group_by_transducer(autd)
         }),
     ];
     if autd.num_devices() >= 2 {
-        examples.push(("Group (by Device) test", |autd| {
-            Box::pin(group_by_device(autd))
-        }));
+        examples.push(("Group (by Device) test", |autd| (group_by_device(autd))));
     }
 
     loop {
@@ -58,7 +53,7 @@ pub async fn run<L: Link>(mut autd: Controller<L>) -> anyhow::Result<()> {
         io::stdin().read_line(&mut s)?;
         match s.trim().parse::<usize>() {
             Ok(i) if i < examples.len() => {
-                if !(examples[i].1)(&mut autd).await? {
+                if !(examples[i].1)(&mut autd)? {
                     eprintln!("Failed to send data");
                 }
             }
@@ -69,11 +64,11 @@ pub async fn run<L: Link>(mut autd: Controller<L>) -> anyhow::Result<()> {
         let mut _s = String::new();
         io::stdin().read_line(&mut _s)?;
 
-        autd.send((Static::new(), Null::default())).await?;
-        autd.send(Silencer::default()).await?;
+        autd.send((Static::new(), Null::default()))?;
+        autd.send(Silencer::default())?;
     }
 
-    autd.close().await?;
+    autd.close()?;
 
     Ok(())
 }