+++ examples/src/twincat.rs
@@ -5,11 +5,8 @@ use anyhow::Result;
 use autd3::prelude::*;
 use autd3_link_twincat::TwinCAT;
 
-#[tokio::main]
-async fn main() -> Result<()> {
-    let autd = Controller::builder([AUTD3::new(Point3::origin())])
-        .open(TwinCAT::builder())
-        .await?;
+fn main() -> Result<()> {
+    let autd = Controller::builder([AUTD3::new(Point3::origin())]).open(TwinCAT::builder())?;
 
-    tests::run(autd).await
+    tests::run(autd)
 }